Added News, Quick Links and Current Location
Couldn't find a way to get local news only, so for now I went with using
Google's top news. I also made it where when you click on the news tile
it pops up with more news. The current location of the user displays at
the top and I took the social media links from the bottom and put them
in a tile box linking them to the homepage of that site for user's quick
access.