Weather Implementation plus Visuals
I couldn't get the full screen weather pictures as I hoped, but I'm not
sure if Andy was going to try to get a photo of the day anyways, so I
added a weather widget that collects weather by location and I changed
the background image just to put something other than a plain background
for now.  We don't have to go with this but I just put something there.
I also added a different font.  Again, not something we have to keep.